refactor: github 충돌 해결 및 간단 리팩터링
