+++ backend/src/feed/feed.controller.ts
@@ -1,6 +1,6 @@
 import { Body, Controller, Get, Param, Patch, Post } from '@nestjs/common';
 
-import Feed from 'src/customDecorator/feed.decorator';
+import Feed from 'src/custom/customDecorator/feed.decorator';
 import ValidationPipe422 from 'src/validation';
 import CreateFeedDto from './dto/create.feed.dto';
 import { FeedService } from './feed.service';
