+++ backend/src/users/dto/join.nickname.dto.ts
@@ -1,4 +1,4 @@
-import IsValidNickname from 'src/customValidators/nicknameValidate';
+import IsValidNickname from 'src/custom/customValidators/nicknameValidate';
 
 export default class JoinNicknameDto {
   @IsValidNickname({ groups: ['customNickname'] })
